requirements.txt
asgiref==3.6.0
Django==4.1.6
django-environ==0.9.0
gunicorn==20.1.0
sqlparse==0.4.3
tzdata==2022.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.8.0
requests>=2.26.0,<3.0
psycopg2-binary<3,>=2.9

